+++ src/components/Inputs.tsx
@@ -24,6 +24,7 @@ import Tags from '@yaireo/tagify/dist/react.tagify'
 import { isEqual } from 'lodash'
 
 import { sectionChildrenVariants } from './PageComponents/PageContainers'
+import Popup from './Popup'
 
 interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
   error?: string
@@ -230,7 +231,7 @@ export function Select<T>({
             options={options}
             setValue={setInputValue}
             title={title}
-            handleBackgroundClick={() => {
+            onBackgroundClick={() => {
               setShowPopup(false)
             }}
           />
@@ -243,48 +244,27 @@ export function Select<T>({
 function SelectOptionsPopup<T>({
   options,
   setValue,
-  handleBackgroundClick,
+  onBackgroundClick,
   title
 }: {
   options: SelectOption<T>[]
   setValue: (value: SelectOption<T>) => void | undefined
-  handleBackgroundClick: () => void
+  onBackgroundClick: () => void
   title?: string
 }) {
   const handleOptionSelect = (value: SelectOption<T>) => {
     setValue(value)
-    handleBackgroundClick()
+    onBackgroundClick()
   }
 
   return (
-    <PopupContainer
-      initial={{ opacity: 0 }}
-      animate={{ opacity: 1 }}
-      exit={{ opacity: 0 }}
-      onClick={() => {
-        handleBackgroundClick()
-      }}
-    >
-      <Popup
-        onClick={(e) => {
-          e.stopPropagation()
-        }}
-        initial={{ y: -10 }}
-        animate={{ y: 0 }}
-        exit={{ y: -10 }}
-      >
-        {title && (
-          <SelectOptionsHeader>
-            <h2>{title}</h2>
-          </SelectOptionsHeader>
-        )}
-        {options.map((o) => (
-          <OptionItem key={o.label} onClick={() => handleOptionSelect(o)}>
-            {o.label}
-          </OptionItem>
-        ))}
-      </Popup>
-    </PopupContainer>
+    <Popup title={title} onBackgroundClick={onBackgroundClick}>
+      {options.map((o) => (
+        <OptionItem key={o.label} onClick={() => handleOptionSelect(o)}>
+          {o.label}
+        </OptionItem>
+      ))}
+    </Popup>
   )
 }
 
@@ -413,29 +393,6 @@ const SelectContainer = styled(InputContainer)`
   cursor: pointer;
 `
 
-const PopupContainer = styled(motion.div)`
-  position: fixed;
-  top: 0;
-  right: 0;
-  left: 0;
-  bottom: 0;
-  display: flex;
-  background-color: rgba(0, 0, 0, 0.6);
-  z-index: 1000;
-`
-
-const Popup = styled(motion.div)`
-  border-radius: var(--radius);
-  margin: auto;
-  width: 30vw;
-  min-width: 300px;
-  max-height: 500px;
-  overflow-x: hidden;
-  overflow-y: auto;
-  box-shadow: var(--shadow);
-  background-color: ${({ theme }) => theme.bg.primary};
-`
-
 const OptionItem = styled.div`
   padding: var(--spacing-3);
   cursor: pointer;
@@ -448,14 +405,6 @@ const OptionItem = styled.div`
   &:hover {
     background-color: ${({ theme }) => theme.bg.secondary};
   }
-`
-
-const SelectOptionsHeader = styled.header`
-  padding: var(--spacing-1) var(--spacing-3);
-  border-bottom: 1px solid ${({ theme }) => theme.border.primary};
-  background-color: ${({ theme }) => theme.bg.secondary};
-  display: flex;
-  align-items: center;
 `
 
 export const Form = styled.form`
